+++ src/cache.d.ts
@@ -1,0 +1,8 @@
+
+interface BasicCachedEvent {
+	created: number;
+	updated: number;
+	ends: number;
+
+	data: any;
+}
